refector : #186 api 디렉토리 구조 변경
