+++ client/src/types/auth.ts
@@ -1,7 +1,12 @@
-interface PostResetPasswordBody {
+export interface PostResetPasswordBody {
   token: string;
   password: string;
   passwordConfirmation: string;
 }
 
-export default PostResetPasswordBody;
+export interface signupProps {
+  email: string;
+  password: string;
+  passwordConfirmation: string;
+  nickname: string;
+}
